refactor: add an asset controller on server
