+++ server/index.js
@@ -13,6 +13,7 @@ require('./models/tech');
 require('./services/passport');
 
 const clientRoutes = require('./routes/clientRoutes');
+const assetRoutes = require('./routes/assetRoutes');
 
 // connect to the database
 const connDb = mongoose.connect(keys.mongoURI).then(
@@ -45,7 +46,7 @@ app.use(morgan('dev'));
 // import routes
 require('./routes/authRoutes')(app);
 app.use('/api/whd/clients', clientRoutes);
-require('./routes/assetRoutes')(app);
+app.use('/api/whd/assets', assetRoutes);
 
 if (process.env.NODE_ENV === 'production') {
   app.use(express.static('client/build'));
